Server/requirements.txt
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.5
drf-spectacular==0.27.0
django-cors-headers==4.3.1
